Rename globe to vertical-perspective
