+++ src/geo/projection/projection_factory.ts
@@ -3,7 +3,7 @@ import {warnOnce} from '../../util/util';
 import {MercatorProjection} from './mercator_projection';
 import {MercatorTransform} from './mercator_transform';
 import {MercatorCameraHelper} from './mercator_camera_helper';
-import {GlobeProjection} from './globe_projection';
+import {VerticalPerspectiveProjection} from './vertial_perspective_projection';
 import {GlobeTransform} from './globe_transform';
 import {GlobeCameraHelper} from './globe_camera_helper';
 import {VerticalPerspectiveTransform} from './vertical_perspective_transform';
@@ -27,7 +27,7 @@ export function createProjectionFromName(name: ProjectionSpecification['type']):
         }
         case 'globe':
         {
-            const proj = new GlobeProjection();
+            const proj = new VerticalPerspectiveProjection();
             return {
                 projection: proj,
                 transform: new GlobeTransform(proj),
@@ -36,7 +36,7 @@ export function createProjectionFromName(name: ProjectionSpecification['type']):
         }
         case 'vertical-perspective':
         {
-            const proj = new GlobeProjection();
+            const proj = new VerticalPerspectiveProjection();
             return {
                 projection: proj,
                 transform: new VerticalPerspectiveTransform(),
